+++ src/app/(main)/market/_layout/Mobile/index.tsx
@@ -1,18 +1,16 @@
 import { PropsWithChildren } from 'react';
-import { Flexbox } from 'react-layout-kit';
+
+import MobileContentLayout from '@/components/server/MobileNavLayout';
 
 import AgentSearchBar from '../../features/AgentSearchBar';
 import Header from './Header';
 
 const MobileLayout = ({ children }: PropsWithChildren) => {
   return (
-    <>
-      <Header />
-      <Flexbox flex={1} gap={16} style={{ padding: 16 }}>
-        <AgentSearchBar mobile />
-        {children}
-      </Flexbox>
-    </>
+    <MobileContentLayout gap={16} header={<Header />} style={{ paddingInline: 16 }} withNav>
+      <AgentSearchBar mobile />
+      {children}
+    </MobileContentLayout>
   );
 };