+++ src/libs/agent-runtime/minimax/index.ts
@@ -49,7 +49,7 @@ function throwIfErrorResponse(data: MinimaxResponse) {
   });
 }
 
-function parseMinimaxResponse(chunk: string): string | undefined {
+function parseMinimaxResponse(chunk: string): MinimaxResponse | undefined {
   let body = chunk;
   if (body.startsWith('data:')) {
     body = body.slice(5).trim();
@@ -57,10 +57,7 @@ function parseMinimaxResponse(chunk: string): string | undefined {
   if (isEmpty(body)) {
     return;
   }
-  const data = JSON.parse(body) as MinimaxResponse;
-  if (data.choices?.at(0)?.delta?.content) {
-    return data.choices.at(0)?.delta.content || undefined;
-  }
+  return JSON.parse(body) as MinimaxResponse;
 }
 
 export class LobeMinimaxAI implements LobeRuntimeAI {
@@ -136,11 +133,25 @@ export class LobeMinimaxAI implements LobeRuntimeAI {
     }
   }
 
+  // the document gives the default value of max tokens, but abab6.5 and abab6.5s
+  // will meet length finished error, and output is truncationed
+  // so here fill the max tokens number to fix it
+  // https://www.minimaxi.com/document/guides/chat-model/V2
+  private getMaxTokens(model: string): number | undefined {
+    switch (model) {
+      case 'abab6.5-chat':
+      case 'abab6.5s-chat': {
+        return 2048;
+      }
+    }
+  }
+
   private buildCompletionsParams(payload: ChatStreamPayload) {
     const { temperature, top_p, ...params } = payload;
 
     return {
       ...params,
+      max_tokens: this.getMaxTokens(payload.model),
       stream: true,
       temperature: temperature === 0 ? undefined : temperature,
       top_p: top_p === 0 ? undefined : top_p,
@@ -159,7 +170,8 @@ export class LobeMinimaxAI implements LobeRuntimeAI {
       const { value, done: doneReading } = await reader.read();
       done = doneReading;
       const chunkValue = decoder.decode(value, { stream: true });
-      const text = parseMinimaxResponse(chunkValue);
+      const data = parseMinimaxResponse(chunkValue);
+      const text = data?.choices?.at(0)?.delta?.content || undefined;
       streamController?.enqueue(encoder.encode(text));
     }
 
@@ -173,12 +185,14 @@ export class LobeMinimaxAI implements LobeRuntimeAI {
     const chunkValue = decoder.decode(value, { stream: true });
     let data;
     try {
-      data = JSON.parse(chunkValue) as MinimaxResponse;
+      data = parseMinimaxResponse(chunkValue);
     } catch {
       // parse error, skip it
       return;
     }
-    throwIfErrorResponse(data);
+    if (data) {
+      throwIfErrorResponse(data);
+    }
   }
 }